frontload appeal/task processing; consolidate tasks; add default args/types to queue/utils
