+++ client/test/node/queue/QueueLoadingScreen-test.js
@@ -1,47 +1,125 @@
 import { expect } from 'chai';
-import { associateTasksWithAppeals } from '../../../app/queue/utils';
+import { associateTasksWithAppeals, mapArrayToObjectById, sortTasks } from '../../../app/queue/utils';
+
+const serverData = {
+  appeals: {
+    data: [{
+      id: '123',
+      attributes: {
+        vacols_id: '1',
+        aod: true
+      }
+    }, {
+      id: '234',
+      attributes: {
+        vacols_id: '2',
+        type: 'Court Remand'
+      }
+    }, {
+      id: '345',
+      attributes: { vacols_id: '3' }
+    }]
+  },
+  tasks: {
+    data: [{
+      id: '111',
+      attributes: {
+        appeal_id: '1',
+        docket_date: '2017-12-28T17:18:20.412Z'
+      }
+    }, {
+      id: '222',
+      attributes: {
+        appeal_id: '1',
+        docket_date: '2016-10-07T03:15:27.580Z'
+      }
+    }, {
+      id: '333',
+      attributes: {
+        appeal_id: '2',
+        docket_date: '2015-10-13T06:47:34.155Z'
+      }
+    }, {
+      id: '444',
+      attributes: {
+        appeal_id: '3',
+        docket_date: '2016-03-01T04:15:51.123Z'
+      }
+    }]
+  }
+};
 
 describe('QueueLoadingScreen', () => {
   it('associates queue decisions/appeals and tasks', () => {
-    const appeals = {
-      data: [
-        { attributes: { vacols_id: '1' } },
-        { attributes: { vacols_id: '2' } }
-      ]
-    };
-    const tasks = {
-      data: [
-        { attributes: { appeal_id: '1' } },
-        { attributes: { appeal_id: '1' } },
-        { attributes: { appeal_id: '2' } }
-      ]
-    };
-    const appealsWithTasks = associateTasksWithAppeals({
-      appeals,
-      tasks
-    });
+    const { tasks: tasksWithAppeals } = associateTasksWithAppeals(serverData);
 
-    expect(appealsWithTasks).to.deep.equal({
-      appeals: [
-        { attributes: { vacols_id: '1' } },
-        { attributes: { vacols_id: '2' } }
-      ],
-      tasks: [{
-        attributes: { appeal_id: '1' },
-        appeal: {
-          attributes: { vacols_id: '1' }
-        }
-      }, {
-        attributes: { appeal_id: '1' },
-        appeal: {
-          attributes: { vacols_id: '1' }
-        }
-      }, {
-        attributes: { appeal_id: '2' },
-        appeal: {
-          attributes: { vacols_id: '2' }
-        }
-      }]
-    });
+    expect(tasksWithAppeals).to.deep.equal([{
+      id: '111',
+      appealId: '123',
+      attributes: {
+        appeal_id: '1',
+        docket_date: '2017-12-28T17:18:20.412Z'
+      },
+    }, {
+      id: '222',
+      appealId: '123',
+      attributes: {
+        appeal_id: '1',
+        docket_date: '2016-10-07T03:15:27.580Z'
+      }
+    }, {
+      id: '333',
+      appealId: '234',
+      attributes: {
+        appeal_id: '2',
+        docket_date: '2015-10-13T06:47:34.155Z'
+      }
+    }, {
+      id: '444',
+      appealId: '345',
+      attributes: {
+        appeal_id: '3',
+        docket_date: '2016-03-01T04:15:51.123Z'
+      }
+    }]);
+  });
+
+  it('groups tasks by AOD/CAVC and sorts by docket date', () => {
+    let { tasks, appeals } = associateTasksWithAppeals(serverData);
+
+    tasks = mapArrayToObjectById(tasks);
+    appeals = mapArrayToObjectById(appeals, { docCount: 0 });
+
+    const sortedTasks = sortTasks({ tasks, appeals });
+
+    expect(sortedTasks).to.deep.equal([{
+      id: '333',
+      appealId: '234',
+      attributes: {
+        appeal_id: '2',
+        docket_date: '2015-10-13T06:47:34.155Z'
+      }
+    }, {
+      id: '222',
+      appealId: '123',
+      attributes: {
+        appeal_id: '1',
+        docket_date: '2016-10-07T03:15:27.580Z'
+      }
+    }, {
+      id: '111',
+      appealId: '123',
+      attributes: {
+        appeal_id: '1',
+        docket_date: '2017-12-28T17:18:20.412Z'
+      }
+    }, {
+      id: '444',
+      appealId: '345',
+      attributes: {
+        appeal_id: '3',
+        docket_date: '2016-03-01T04:15:51.123Z'
+      }
+    }]);
   });
 });
